+++ app/src/main/java/com/mangoboss/app/api/facade/task/StaffTaskFacade.java
@@ -1,7 +1,6 @@
 package com.mangoboss.app.api.facade.task;
 
-import com.mangoboss.app.common.constant.S3FileType;
-import com.mangoboss.app.common.util.S3FileManager;
+import com.mangoboss.app.common.util.S3PreSignedUrlManager;
 import com.mangoboss.app.domain.service.staff.StaffService;
 import com.mangoboss.app.domain.service.task.TaskService;
 import com.mangoboss.app.dto.s3.response.UploadPreSignedUrlResponse;
@@ -9,6 +8,7 @@ import com.mangoboss.app.dto.staff.response.StaffSimpleResponse;
 import com.mangoboss.app.dto.task.request.TaskCheckRequest;
 import com.mangoboss.app.dto.task.response.AssignedTaskResponse;
 import com.mangoboss.app.dto.task.response.TaskLogDetailResponse;
+import com.mangoboss.storage.metadata.S3FileType;
 import com.mangoboss.storage.task.TaskEntity;
 import com.mangoboss.storage.task.TaskLogEntity;
 import lombok.RequiredArgsConstructor;
@@ -25,7 +25,7 @@ public class StaffTaskFacade {
 
     private final TaskService taskService;
     private final StaffService staffService;
-    private final S3FileManager s3FileManager;
+    private final S3PreSignedUrlManager s3PreSignedUrlManager;
 
     public void completeTask(final Long storeId, final Long taskId, final Long userId, final TaskCheckRequest request) {
         staffService.getVerifiedStaff(userId, storeId);
@@ -40,8 +40,8 @@ public class StaffTaskFacade {
     public UploadPreSignedUrlResponse generateReportImageUploadUrl(final Long storeId, final Long userId,
                                                                    final String extension, final String contentType) {
         staffService.getVerifiedStaff(userId, storeId);
-        final String key = s3FileManager.generateFileKey(S3FileType.TASK_REPORT, extension);
-        return s3FileManager.generateUploadPreSignedUrl(key, contentType);
+        final String key = s3PreSignedUrlManager.generateFileKey(S3FileType.TASK_REPORT, extension);
+        return s3PreSignedUrlManager.generateUploadPreSignedUrl(key, contentType);
     }
 
     public List<AssignedTaskResponse> getTasksByDate(final Long storeId, final Long userId, final LocalDate date) {
