+++ app/src/main/java/com/mangoboss/app/common/util/S3FileManager.java
@@ -1,24 +1,16 @@
 package com.mangoboss.app.common.util;
 
-import com.mangoboss.app.common.constant.ContentType;
-import com.mangoboss.app.common.constant.S3FileType;
+import com.mangoboss.app.common.config.S3Config;
 import com.mangoboss.app.common.exception.CustomErrorInfo;
 import com.mangoboss.app.common.exception.CustomException;
-import com.mangoboss.app.dto.s3.response.DownloadPreSignedUrlResponse;
-import com.mangoboss.app.dto.s3.response.UploadPreSignedUrlResponse;
-import com.mangoboss.app.dto.s3.response.ViewPreSignedUrlResponse;
-import lombok.Getter;
+import com.mangoboss.storage.metadata.S3FileType;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import software.amazon.awssdk.core.ResponseBytes;
 import software.amazon.awssdk.core.sync.RequestBody;
 import software.amazon.awssdk.services.s3.S3Client;
 import software.amazon.awssdk.services.s3.model.*;
-import software.amazon.awssdk.services.s3.presigner.S3Presigner;
-import software.amazon.awssdk.services.s3.presigner.model.*;
 
-import java.time.*;
 import java.util.Base64;
 import java.util.UUID;
 
@@ -28,64 +20,39 @@ public class S3FileManager {
     private static final String SSE_ALGORITHM = "aws:kms";
 
     private final S3Client s3Client;
-    private final S3Presigner s3Presigner;
-    private final Clock clock;
-
-    @Getter
-    @Value("${cloud.aws.s3.public-base-url}")
-    private String publicBaseUrl;
-
-    @Value("${cloud.aws.s3.kms-key-id}")
-    private String kmsKeyId;
-
-    @Value("${pre-signed-url.view-expiration-minutes}")
-    private int viewExpirationMinutes;
-
-    @Value("${pre-signed-url.download-expiration-minutes}")
-    private int downloadExpirationMinutes;
-
-    @Value("${pre-signed-url.upload-expiration-minutes}")
-    private int uploadExpirationMinutes;
-
-    @Getter
-    @Value("${cloud.aws.s3.private-bucket}")
-    private String privateBucket;
-
-    @Getter
-    @Value("${cloud.aws.s3.public-bucket}")
-    private String publicBucket;
-
+    private final S3Config s3Config;
 
     // 파일 업로드: byte[] 기반
     public void upload(final byte[] fileData, final String key, final String contentType) {
         final PutObjectRequest request = PutObjectRequest.builder()
-                .bucket(privateBucket)
+                .bucket(s3Config.getPrivateBucket())
                 .key(key)
                 .contentType(contentType)
                 .serverSideEncryption(SSE_ALGORITHM)
-                .ssekmsKeyId(kmsKeyId)
+                .ssekmsKeyId(s3Config.getKmsKeyId())
                 .build();
 
         s3Client.putObject(request, RequestBody.fromBytes(fileData));
     }
 
-    public void upload(String key, byte[] fileBytes, ContentType contentType) {
-        final PutObjectRequest request = PutObjectRequest.builder()
-                .bucket(privateBucket)
-                .key(key)
-                .contentType(contentType.getMimeType())
-                .build();
-
-        s3Client.putObject(request, RequestBody.fromBytes(fileBytes));
-    }
-
-    public void deleteFile(final String key) {
+    public void deleteFileFromPrivateBucket(final String key) {
         try {
             final DeleteObjectRequest deleteRequest = DeleteObjectRequest.builder()
-                    .bucket(privateBucket)
+                    .bucket(s3Config.getPrivateBucket())
                     .key(key)
                     .build();
+            s3Client.deleteObject(deleteRequest);
+        } catch (Exception e) {
+            throw new CustomException(CustomErrorInfo.FILE_DELETE_FAILED);
+        }
+    }
 
+    public void deleteFileFromPublicBucket(final String key) {
+        try {
+            final DeleteObjectRequest deleteRequest = DeleteObjectRequest.builder()
+                    .bucket(s3Config.getPublicBucket())
+                    .key(key)
+                    .build();
             s3Client.deleteObject(deleteRequest);
         } catch (Exception e) {
             throw new CustomException(CustomErrorInfo.FILE_DELETE_FAILED);
@@ -104,7 +71,7 @@ public class S3FileManager {
         try {
             final ResponseBytes<GetObjectResponse> objectBytes = s3Client.getObjectAsBytes(
                     GetObjectRequest.builder()
-                            .bucket(privateBucket)
+                            .bucket(s3Config.getPrivateBucket())
                             .key(key)
                             .build()
             );
@@ -114,92 +81,13 @@ public class S3FileManager {
         }
     }
 
-    // 조회용 presingedUrl (inline)
-    public ViewPreSignedUrlResponse generateViewPreSignedUrl(final String key) {
-        final GetObjectRequest getObjectRequest = GetObjectRequest.builder()
-                .bucket(privateBucket)
-                .key(key)
-                .responseContentDisposition("inline")
-                .build();
-
-        final GetObjectPresignRequest preSignRequest = GetObjectPresignRequest.builder()
-                .signatureDuration(Duration.ofMinutes(viewExpirationMinutes))
-                .getObjectRequest(getObjectRequest)
-                .build();
-
-        final String url = s3Presigner.presignGetObject(preSignRequest).url().toString();
-
-        return ViewPreSignedUrlResponse.of(url, LocalDateTime.now(clock).plusMinutes(viewExpirationMinutes));
-    }
-
-    // 다운로드용 presingedUrl
-    public DownloadPreSignedUrlResponse generateDownloadPreSignedUrl(final String key) {
-        final GetObjectRequest getObjectRequest = GetObjectRequest.builder()
-                .bucket(privateBucket)
-                .key(key)
-                .responseContentDisposition("attachment")
-                .build();
-
-        final GetObjectPresignRequest preSignRequest = GetObjectPresignRequest.builder()
-                .signatureDuration(Duration.ofMinutes(downloadExpirationMinutes))
-                .getObjectRequest(getObjectRequest)
-                .build();
-
-        final String url = s3Presigner.presignGetObject(preSignRequest).url().toString();
-
-        return DownloadPreSignedUrlResponse.of(url, LocalDateTime.now(clock).plusMinutes(downloadExpirationMinutes));
-    }
-
-    // 업로드용 presingedUrl
-    public UploadPreSignedUrlResponse generateUploadPreSignedUrl(final String key, final String contentType) {
-        final PutObjectRequest putObjectRequest = PutObjectRequest.builder()
-                .bucket(publicBucket)
-                .key(key)
-                .contentType(contentType)
-                .build();
-
-        final PutObjectPresignRequest preSignRequest = PutObjectPresignRequest.builder()
-                .signatureDuration(Duration.ofMinutes(uploadExpirationMinutes))
-                .putObjectRequest(putObjectRequest)
-                .build();
-
-        final String uploadUrl = s3Presigner.presignPutObject(preSignRequest).url().toString();
-        final String publicUrl = publicBaseUrl + key;
-
-        return UploadPreSignedUrlResponse.of(uploadUrl, publicUrl, LocalDateTime.now(clock).plusMinutes(uploadExpirationMinutes));
-    }
-
-    public void deleteFileFromTaskBucket(final String key) {
-        try {
-            final DeleteObjectRequest deleteRequest = DeleteObjectRequest.builder()
-                    .bucket(publicBucket)
-                    .key(key)
-                    .build();
-            s3Client.deleteObject(deleteRequest);
-        } catch (Exception e) {
-            throw new CustomException(CustomErrorInfo.FILE_DELETE_FAILED);
-        }
-    }
-
     public String generateFileKey(final S3FileType fileType, String extension) {
         return fileType.getFolder() + UUID.randomUUID() + "." + extension;
     }
 
-    public String getContentTypeFromS3(final String key) {
-        try {
-            final HeadObjectRequest request = HeadObjectRequest.builder()
-                    .bucket(privateBucket)
-                    .key(key)
-                    .build();
+    public String extractKeyFromPublicUrl(final String url) {
+        final String baseUrl = s3Config.getPublicBaseUrl();
 
-            final HeadObjectResponse response = s3Client.headObject(request);
-            return response.contentType();
-        } catch (Exception e) {
-            throw new CustomException(CustomErrorInfo.S3_OBJECT_FETCH_FAILED);
-        }
-    }
-
-    public String extractKeyFromUrl(final String url, final String baseUrl) {
         if (!url.startsWith(baseUrl)) {
             throw new CustomException(CustomErrorInfo.INVALID_S3_URL);
         }
