+++ storage/src/main/java/com/mangoboss/storage/metadata/S3FileType.java
@@ -9,7 +9,9 @@ public enum S3FileType {
     CONTRACT("contracts/"),
     SIGNATURE("signatures/"),
     DOCUMENT("documents/"),
-    PAYSLIP("payslips/");
+    PAYSLIP("payslips/"),
+    TASK_REFERENCE("tasks/reference/"),
+    TASK_REPORT("tasks/report/");
 
     private final String folder;
 }
